refactor: posts files into posts folder, imports are up to date
