+++ client/src/components/feed/Feed.js
@@ -10,7 +10,7 @@ const [{user}, dispatch] = useStateValue();
 //posts array
 const [posts, setPosts] = useState([]);
 //lazy loading
-const Post = React.lazy(() => import('../Post'))
+const Post = React.lazy(() => import('../posts/Post'))
 //====================================Post changes listner=========================================
     useEffect( () => {
         //onSnapshot = listner to changes in posts
